🎨 Make icons individual components
