+++ components/Navbar.jsx
@@ -1,8 +1,9 @@
-import React, { useState, Fragment, useEffect, useRef } from "react";
+import React, { useState, Fragment } from "react";
 import { Menu, Transition } from "@headlessui/react";
 import Link from "next/link";
 import { ChevronDownIcon } from "@heroicons/react/solid";
 import { SiDiscord } from "react-icons/si";
+import { ArchiveIcon, EditIcon, DuplicateIcon, DeleteIcon } from "./Icons";
 
 import {
   AiOutlineMenuFold,
@@ -414,138 +415,3 @@ const Navbar = (props) => {
 };
 
 export default Navbar;
-
-function EditIcon({ active }) {
-  return (
-    <svg
-      className="mr-2 h-5 w-5 text-violet-400"
-      aria-hidden="true"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M4 13V16H7L16 7L13 4L4 13Z"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-    </svg>
-  );
-}
-
-function DuplicateIcon({ active }) {
-  return (
-    <svg
-      className="mr-2 h-5 w-5 text-violet-400"
-      aria-hidden="true"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M4 4H12V12H4V4Z"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke="#C4B5FD"
-        strokeWidth="2"
-      />
-      <path
-        d="M8 8H16V16H8V8Z"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-    </svg>
-  );
-}
-
-function ArchiveIcon({ active }) {
-  return (
-    <svg
-      className="mr-2 h-5 w-5"
-      aria-hidden="true"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect
-        x="5"
-        y="8"
-        width="10"
-        height="8"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <rect
-        x="4"
-        y="4"
-        width="12"
-        height="4"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <path d="M8 12H12" stroke="#A78BFA" strokeWidth="2" />
-    </svg>
-  );
-}
-
-function MoveIcon({ active }) {
-  return (
-    <svg
-      {...props}
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M10 4H16V10"
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <path
-        d="M16 4L8 12"
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <path
-        d="M8 6H4V16H14V12"
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-    </svg>
-  );
-}
-
-function DeleteIcon({ active }) {
-  return (
-    <svg
-      className="mr-2 h-5 w-5"
-      aria-hidden="true"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect
-        x="5"
-        y="6"
-        width="10"
-        height="10"
-        fill={active ? "#8B5CF6" : "#EDE9FE"}
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <path
-        d="M3 6H17"
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-      <path
-        d="M8 6V4H12V6"
-        stroke={active ? "#C4B5FD" : "#A78BFA"}
-        strokeWidth="2"
-      />
-    </svg>
-  );
-}
